Improve ManualReleaseNativeContext semantics
