+++ realm/realm-library/src/main/java/io/realm/internal/NativeContext.java
@@ -17,6 +17,7 @@
 package io.realm.internal;
 
 import java.lang.ref.ReferenceQueue;
+import java.util.LinkedList;
 
 
 // Currently we free native objects in two threads, the SharedGroup is freed in the caller thread, others are freed in
@@ -39,5 +40,38 @@ public class NativeContext {
 
     public void addReference(NativeObject referent) {
         new NativeObjectReference(this, referent, referenceQueue);
+    }
+
+    /**
+     * Executes a given NativeContextRunnable with a NativeContext which lifecycle is bound
+     * to the runnable scope.
+     *
+     * @param runnable function to execute.
+     */
+    static void execute(ManualReleaseNativeContext.NativeContextRunnable runnable) {
+        ManualReleaseNativeContext nativeContext = new ManualReleaseNativeContext();
+        runnable.run(nativeContext);
+        nativeContext.release();
+    }
+
+    public interface NativeContextRunnable {
+        void run(NativeContext nativeContext);
+    }
+
+    private static class ManualReleaseNativeContext extends NativeContext {
+        private final LinkedList<NativeObject> references = new LinkedList<>();
+
+        ManualReleaseNativeContext(){ }
+
+        @Override
+        public void addReference(NativeObject referent) {
+            references.add(referent);
+        }
+
+        public void release() {
+            for (NativeObject object : references) {
+                NativeObjectReference.nativeCleanUp(object.getNativeFinalizerPtr(), object.getNativePtr());
+            }
+        }
     }
 }
